DCPU16.Assembly.Optimizer

Description

Simple assembly optimizations.

Focus on small tweaks, like shortening instructions with small literals, and any NOP/call/arithmetic optimizations that come up.

Documentation

data Optimization

Optimizations, listed in the order they're performed. (Earliest first.)

Constructors

ShortLiterals
Shortens instructions by 1 word if they use literals smaller than 0x20.

 set a,1            ; 1 word, not 2
 add [0x8000],1     ; 2 words, not 3

ShortLabelLiterals
Performs the short literal optimization on labels.

 :l_0000    set pc,l.0000       ; 1 word, address < 0x20
 :l_8000    set pc,l.8000       ; 2 words, large address

sizeVariant :: Step

Optimizations that change the size of instructions, but not order.

Should be run before label-to-address translation.