DCPU16.Assembly.Parser

Description

Parser for what seems to be DCPU16 assembly.

There is some ambiguity between sources: the specification uses uppercase a lot (which I'd rather put in as an option later, with the strict implementation being default).

A screenshot also shows indirect mode being done with [] instead of (). Go figure.

Documentation

parseFile :: Options -> String -> IO (Maybe (Vector Instruction))

defaults :: Options

Default parsing options.

data Options

Parsing options, if you want to override the defaults.

Constructors

Options

Fields

allowUppercase :: Bool

Parse upper case symbols. Default: off.

Large amounts of assembly get written in lower case. Upper case might look pretty in small amounts, but holding down the shift key gets old.

roundedBrackets :: Bool

Indirect mode via () instead of []. Default: off.

Weird, but showed up in a screenshot.