DCPU16.Disassembler

Description

Disassembler.

I should really convert these to use Vector Word16, rather than ByteString.

It won't be useful for command line tools, but make more sense for everything else.

High level disassembly

disassemble :: ByteString -> Vector Instruction

Disassembler.

Annotates result with labels for obvious jumps/calls, but keeps operand addresses as literals.

Component functions

readInstructions :: ByteString -> Vector Instruction

Naive disassembly.

Disassembly can fail at the very end of a buffer, if a partial multi-word instruction is encountered. I'll eventually treat those as data words instead, but for now it crashes... And is a point of weakness to keep in mind for interpreters. What will the official one do?

It can also produce gibberish if data words are inserted that look like multi-word instructions. This can be hard to catch, and will throw naive disassemblers.

annotate :: Vector Instruction -> Vector Instruction

Add easily visible labels to instructions.

Stuff like JSR foo SET PC, bar SUB PC, 5

Labels are not inserted as operands, only defined. The names have the address in them, to match up with operands.