DCPU16.Instructions.Time

Description

DCPU-16 instruction execution time in cycles.

Documentation

time :: Instruction -> Int

Instruction execution time in cycles.

Depends on the instruction, and type of operands used.

There's also a 1-cycle cost for unsuccessful IF? comparisons, but that can't be predicted without running the code. As such, this prediction is optimistic and assumes all comparisons succeed.

Since the most cycle-efficient way to write loops, is to jump back on a successful comparison, this prediction should be close to accurate.