DCPU16.Assembly.Optimizer

Description

Simple assembly optimizations.

Focus on small tweaks, like shortening instructions with small literals, and any NOP/call/arithmetic optimizations that come up.

Documentation

sizeVariant :: Vector Instruction -> Vector Instruction

Optimizations that change the size of instructions, but not order.

Should be run before label-to-address translation.